Further improved bomb timing routine 94% solution
This version has a less hasty bomb decision routine (although it is adjusted to always place a bomb on the first turn). This generally gives improved performance, however it still fails test 12 "Not so fast", which is designed to fail 'greedy' solutions such as this. I'm not clear what to do about this. The problem is optimal bomb placement given a limited number of bombs. A greedy solution is not always correct. Perhaps the greedy placement could be alerted to this, by calculating the number of 'distance-separated' nodes that will remain and the number of bombs that will remain. If the former exceed that latter, then the solution should ignore this placement and try the next best. Calculated 'distance-separated' (can't both be hit by same bomb) is quite tricky. They must have either an x value that is at least '2' different from all other nodes, or a y value that is. So we need an "isThisNodeDistant" boolean returning method, that has to ignore nodes that will be threatened by the initial bomb placement. This suggests we need to create a temporary "future" version of the grid to check. We must do all this immediately before selecting the bestPlacement node. and be able to loop to try again if the placement fails the test.